YARN-11672. Create a CgroupHandler implementation for cgroup v2
